requirements.txt.txt
discord.py==2.3.2
paramiko==3.4.0
pytz==2024.2